refactor(client): move sales slice into features/sales
