+++ client/src/app/App.tsx
@@ -15,7 +15,7 @@ import Inventory from '../pages/Inventory';
 // Lazy loaded
 const BarcodeTools = lazy(() => import('../pages/BarcodeTools'));
 const Deliveries = lazy(() => import('../pages/Deliveries'));
-const SalesHistory = lazy(() => import('../pages/SalesHistory'));
+const SalesHistory = lazy(() => import('../features/sales/pages/SalesHistory'));
 const UsersPage = lazy(() => import('../features/admin/pages/Users'));
 const CustomersPage = lazy(() => import('../features/customers/pages/Customers'));
 const DistributorsPage = lazy(() => import('../pages/Distributors'));
@@ -23,8 +23,8 @@ const CategoriesPage = lazy(() => import('../pages/Categories'));
 const SettingsPage = lazy(() => import('../features/admin/pages/Settings'));
 const PurchaseOrdersPage = lazy(() => import('../pages/PurchaseOrders'));
 const AuditLogPage = lazy(() => import('../features/admin/pages/AuditLog'));
-const PromotionsPage = lazy(() => import('../pages/Promotions'));
-const GiftCardsPage = lazy(() => import('../pages/GiftCards'));
+const PromotionsPage = lazy(() => import('../features/sales/pages/Promotions'));
+const GiftCardsPage = lazy(() => import('../features/sales/pages/GiftCards'));
 const StockCountPage = lazy(() => import('../pages/StockCount'));
 const ExportsPage = lazy(() => import('../pages/Exports'));
 const CustomerDisplay = lazy(() => import('../pages/CustomerDisplay'));
@@ -32,7 +32,7 @@ const RegisterPage = lazy(() => import('../pages/Register'));
 const ShiftsPage = lazy(() => import('../pages/Shifts'));
 const ExpensesPage = lazy(() => import('../pages/Expenses'));
 const SegmentsPage = lazy(() => import('../features/customers/pages/Segments'));
-const LayawayPage = lazy(() => import('../pages/Layaway'));
+const LayawayPage = lazy(() => import('../features/sales/pages/Layaway'));
 const BundlesPage = lazy(() => import('../pages/Bundles'));
 const FeedbackPage = lazy(() => import('../features/customers/pages/Feedback'));
 const BackupPage = lazy(() => import('../features/admin/pages/Backup'));
