+++ client/src/features/sales/components/RefundDialog.tsx
@@ -7,14 +7,20 @@ import {
   DialogDescription,
   DialogHeader,
   DialogTitle,
-} from '../shared/ui/dialog';
-import { Button } from '../shared/ui/button';
-import { Label } from '../shared/ui/label';
-import { Checkbox } from '../shared/ui/checkbox';
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '../shared/ui/select';
-import { formatCurrency } from '../shared/lib/utils';
-import { resource } from '../shared/lib/resource';
-import { useTranslation } from '../shared/i18n/index';
+} from '../../../shared/ui/dialog';
+import { Button } from '../../../shared/ui/button';
+import { Label } from '../../../shared/ui/label';
+import { Checkbox } from '../../../shared/ui/checkbox';
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from '../../../shared/ui/select';
+import { formatCurrency } from '../../../shared/lib/utils';
+import { resource } from '../../../shared/lib/resource';
+import { useTranslation } from '../../../shared/i18n/index';
 import type { SaleItem } from '@/types';
 
 /** Only the refund sub-action is reached from here, so no row shape surfaces. */
